XDbot/pyproject.toml
[tool.poetry]
name = "XDbot"
version = "2.4.7"
description = "Multifunctional QQ robot."
authors = ["This is XiaoDeng", "XieXiLin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
